requirements.txt
aiogram>=3.0,<4
python-dotenv>=1.0,<2
